Respect size_t return value from strlen().
Fixes VS 2005 warning about conversion from size_t to int and
possible loss of data.

+++ jbig2dec/jbig2dec.c
@@ -748,7 +748,7 @@ main(int argc, char **argv)
                     goto cleanup;
                 }
             } else {
-                int len = strlen(params.output_filename);
+                size_t len = strlen(params.output_filename);
 
                 if ((len >= 3) && (params.output_format == jbig2dec_format_none))
                     /* try to set the output type by the given extension */
